requirements.txt
pywin32
pycryptodome
Flask
gunicorn
